Split api/index.ts into single-purpose modules, fixed jscpd
Splits the 1,375-line api/index.ts into one-purpose modules
(booking, folded-booking, packages, listings, public-listing, guards,
helpers, cors), leaving index.ts as a thin route-map wiring file.

The split surfaced eight jscpd clones that had been silently passing
inside the monolith; fixed them by sharing one mechanism for each:
  - withSlugLoaded (helpers.ts): the load-or-respond routing for every
    :slug route, so withActiveListing and the new withPackageContext
    expose one SlugRouteHandler contract instead of each re-spelling
    (request, { slug }, server)
  - FoldedBookingInput + foldChildrenOrError + validateFoldedFields
    (folded-booking.ts): the fold-then-respond and validate-merged-fields
    blocks shared by the parent and package API booking flows
  - applyChildSelectionsToForm takes ctx: the parent's child-lookup map
    is built once inside the helper instead of at each call site
  - loadBookablePackages (discovery.ts): the bookable-package filter
    shared by the /api/listings response and the feeds

Also removes the biome.json max-lines override for the now-small
index.ts.
